chore: move takeEventToBatches into events folder
