+++ web-registry/src/lib/events/takeEventToBatches.ts
